requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cloudevents==1.2.0
deprecation==2.1.0
Flask==2.0.3
functions-framework==3.0.0
google-api-core==2.7.1
google-auth==2.6.2
google-cloud-core==2.2.3
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
kubernetes==23.3.0
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
watchdog==1.0.2
websocket-client==1.3.1
Werkzeug==2.0.3
